module7-task1 начало store
