+++ src/components/city-places-form-filters/city-places-form-filters.jsx
@@ -3,7 +3,6 @@ import PropTypes from "prop-types";
 import {connect} from "react-redux";
 import {sortCityFilterType, sortPlacesInCity} from "../../actions";
 import CityPlacesFormFiltersItem from "../city-places-form-filters-item/city-places-form-filters-item";
-import {selectCurrentPlacesFilter} from "../../reducer/places-filter/selectors";
 
 const CityPlacesFormFilters = (props) => {
   const {onFilterClick, currentPlaceFilterType, isFilterVisible, onSortPlacesInCity} = props;
@@ -38,7 +37,7 @@ CityPlacesFormFilters.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  currentPlaceFilterType: selectCurrentPlacesFilter(state)
+  currentPlaceFilterType: null
 });
 
 const mapDispatchToProps = (dispatch) => ({
