+++ src/components/city-places/city-places.jsx
@@ -4,7 +4,6 @@ import {connect} from "react-redux";
 import PlacesList from "../places-list/places-list";
 import CityPlacesForm from "../city-places-form/city-places-form";
 import withToggleElement from "../../hocs/with-toggle-element/with-toggle-element";
-import {selectCurrentCity} from "../../reducer/city-selected/selectors";
 
 const CityPlacesFormWrapped = withToggleElement(CityPlacesForm);
 
@@ -29,7 +28,7 @@ CityPlaces.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  city: selectCurrentCity(state)
+  city: {}
 });
 
 export {CityPlaces};
