+++ src/components/main/main.jsx
@@ -4,8 +4,6 @@ import {connect} from "react-redux";
 import CityPlaces from "../city-places/city-places";
 import Map from "../map/map";
 import CitiesList from "../cities-list/cities-list";
-import {selectCurrentCityCord} from "../../reducer/city-selected/selectors";
-import {selectPlacesInCity} from "../../reducer/places-in-city/selectors";
 
 const Main = (props) => {
   const {onClickCardTitle, offers, cityCord, mouseOverPlaceCard} = props;
@@ -37,8 +35,8 @@ Main.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  cityCord: selectCurrentCityCord(state),
-  offers: selectPlacesInCity(state),
+  cityCord: [],
+  offers: [],
   mouseOverPlaceCard: state.mouseOverPlaceCard
 });
