--- src/reducer/city-list/actions.js
@@ -1,11 +1,0 @@
-const ActionType = {
-  LOAD_CITY_LIST: `LOAD_CITY_LIST`
-};
-
-const ActionCreator = {
-  loadCityList(payload) {
-    return ({type: ActionType.LOAD_CITY_LIST, payload});
-  }
-};
-
-export {ActionCreator, ActionType};
