--- src/reducer/city-list/selectors.js
@@ -1,7 +1,0 @@
-import StoreNameSpace from "../store-name-space";
-
-const selectCityList = (state) => {
-  return state[StoreNameSpace.CITY_LIST].items;
-};
-
-export {selectCityList};
