--- src/reducer/data/actions.js
@@ -1,11 +1,0 @@
-const ActionType = {
-  LOAD_DATA: `LOAD_DATA`
-};
-
-const ActionCreator = {
-  loadData(payload) {
-    return {type: ActionType.LOAD_DATA, payload}
-  }
-};
-
-export {ActionType, ActionCreator};
