--- src/reducer/data/model.js
@@ -1,8 +1,0 @@
-const serverDataToFrontData = (data) => {
-  return data.map((it) => {
-    it.city.id = it.city.name;
-    return it;
-  });
-};
-
-export {serverDataToFrontData};
