--- src/reducer/data/selectors.js
@@ -1,7 +1,0 @@
-import StoreNameSpace from "../store-name-space";
-
-const getData = (state) => {
-  return state[StoreNameSpace.DATA].data;
-};
-
-export {getData};
