--- src/reducer/places-filter/places-filter.js
@@ -1,11 +1,0 @@
-import {sortCityFilterType} from "../../const";
-
-const initialState = {
-  name: sortCityFilterType.TOP_RATED_FIRST
-};
-
-const reducer = (state = initialState, action) => {
-  return state;
-};
-
-export {reducer};
