--- src/reducer/places-filter/selectors.js
@@ -1,7 +1,0 @@
-import StoreNameSpace from "../store-name-space";
-
-const selectCurrentPlacesFilter = (state) => {
-  return state[StoreNameSpace.PLACES_FILTER].name;
-};
-
-export {selectCurrentPlacesFilter};
