--- src/reducer/places-in-city/selectors.js
@@ -1,8 +1,0 @@
-import StoreNameSpace from "../store-name-space";
-
-const selectPlacesInCity = (state) => {
-  return state[StoreNameSpace.PLACES_IN_CITY].items;
-};
-
-export {selectPlacesInCity};
-
