--- src/reducer/store-name-space.js
@@ -1,9 +1,0 @@
-const StoreNameSpace = {
-  DATA: `DATA`,
-  CITY_LIST: `CITY_LIST`,
-  CITY_SELECTED: `CITY_SELECTED`,
-  PLACES_IN_CITY: `PLACES_IN_CITY`,
-  PLACES_FILTER: `PLACES_FILTER`
-};
-
-export default StoreNameSpace;
