+++ src/store/all-hotels/actions.js
@@ -1,0 +1,11 @@
+const ActionType = {
+  LOAD_ALL_HOTELS: `LOAD_ALL_HOTELS`
+};
+
+const ActionCreator = {
+  loadAllHotels(hotels) {
+    return {type: ActionType.LOAD_ALL_HOTELS, payload: hotels};
+  }
+};
+
+export {ActionType, ActionCreator};
