+++ src/store/cities/reducers.js
@@ -1,8 +1,4 @@
-const initialState = {
-  id: 17,
-  name: `Амстердам`,
-  cord: [52.38333, 4.9]
-};
+const initialState = {};
 
 const reducer = (state = initialState, action) => {
   return state;
