+++ src/store/store-name-space.js
@@ -1,0 +1,7 @@
+const StoreNameSpace = {
+  ALL_HOTELS: `ALL_HOTELS`,
+  CITIES: `CITIES`,
+  PLACES_IN_CITY: `PLACES_IN_CITY`
+};
+
+export default StoreNameSpace;
